Refactor SearchFiltersQuery eval methods
