+++ stac-example/src/main/resources/application.conf
@@ -45,9 +45,14 @@ wms = {
         ${layers.us-ned-hillshade},
         ${layers.us-census-median-household-income},
         ${layers.markham},
-        ${layers.stac-hillshade-buffered},
-        ${layers.stac-hillshade-source},
-        ${layers.stac-hillshade-source-maml}
+        ${layers.stac-lc8-red-us},
+        ${layers.stac-lc8-green-us},
+        ${layers.stac-lc8-blue-us},
+        ${layers.stac-lc8-rgb-us},
+        ${layers.stac-lc8-red-pa},
+        ${layers.stac-lc8-green-pa},
+        ${layers.stac-lc8-blue-pa},
+        ${layers.stac-lc8-rgb-pa}
     ]
 }
 
@@ -72,10 +77,15 @@ wcs = {
         ${layers.us-ned-hillshade},
         ${layers.us-census-median-household-income},
         ${layers.markham},
-        ${layers.stac-hillshade-buffered},
-        ${layers.stac-hillshade-source},
-        ${layers.stac-hillshade-source-maml}
-    ],
+        ${layers.stac-lc8-red-us},
+        ${layers.stac-lc8-green-us},
+        ${layers.stac-lc8-blue-us},
+        ${layers.stac-lc8-rgb-us},
+        ${layers.stac-lc8-red-pa},
+        ${layers.stac-lc8-green-pa},
+        ${layers.stac-lc8-blue-pa},
+        ${layers.stac-lc8-rgb-pa}
+    ]
     supported-projections = [
         4326,
         3410,
@@ -219,81 +229,23 @@ wmts = {
 }
 
 layers = {
-    stac-hillshade-buffered = {
+    stac-lc8-red-us = {
         type = "stacsourceconf"
-        with-gdal = true
-        name = "hillshade-buffered"
-        title = "Hillshade Buffered GeoTrellis Collection."
-        layer = "hillshade-buffered"
-        asset = "cog"
+        name = "stac-lc8-red-us"
+        title = "Landsat LayerUS Red"
+        layer = "layer-us"
+        asset = "B4"
         asset-limit = 1000
         source = "http://localhost:9090/"
         default-time = true
-        default-style = "greyscale"
+        default-style = "red-to-blue"
         styles = [
             {
-                name = "greyscale"
-                title = "Greyscale"
+                name = "red-to-blue"
+                title = "Red To Blue"
                 type = "colorrampconf"
-                colors = ${color-ramps.greyscale}
-                stops = 255,
-                clamp-with-color = true
-            }
-        ]
-    }
-
-    stac-hillshade-source = {
-        type = "stacsourceconf"
-        with-gdal = true
-        name = "stac-hillshade-source"
-        title = "Hillshade Buffered Source Collection."
-        layer = "hillshade-buffered"
-        asset = "source"
-        asset-limit = 1000
-        source = "http://localhost:9090/"
-        default-time = true
-        default-style = "greyscale"
-        styles = [
-            {
-                name = "greyscale"
-                title = "Greyscale"
-                type = "colorrampconf"
-                colors = ${color-ramps.greyscale}
-                stops = 255,
-                clamp-with-color = true
-            }
-        ]
-    }
-
-    stac-hillshade-source-maml = {
-        type = "mapalgebrasourceconf"
-        with-gdal = true
-        name = "hillshade-buffered-source-maml"
-        title = "Hillshade Buffered Source MAML Collection."
-        resample-method = "bilinear"
-        overview-strategy = "auto-0"
-        algebra = {
-            "args" : [
-                {
-                    "name" : "stac-hillshade-source",
-                    "symbol" : "rasterV"
-                }
-            ],
-            "azimuth" : 315.0,
-            "altitude" : 45.0,
-            "zFactor": 1.0,
-            "target" : "all",
-            "symbol" : "fhillshade"
-        }
-        default-style = "greyscale"
-        styles = [
-            {
-                name = "greyscale"
-                title = "Greyscale"
-                type = "colorrampconf"
-                colors = ${color-ramps.greyscale}
-                stops = 255,
-                clamp-with-color = true
+                colors = ${color-ramps.red-to-blue}
+                stops = 64
             }
         ]
     }
@@ -920,12 +872,6 @@ color-ramps = {
         0xB49B55FF, 0x9B7D37FF, 0x6E3C00FF, 0xA59678FF,
         0xD7D7D7FF, 0xFFFFFFFF
     ]
-    "greyscale": [
-        0x000000FF, 0xFFFFFFFF
-    ]
-    "slope": [
-        0x000000FF, 0xFFFFFFFF
-    ]
 }
 
 color-maps = {
@@ -1012,22 +958,5 @@ interpolated-color-maps = {
             "200000.0": 0x00FF00FF,
         }
         "clip-definition": "clip-both"
-    }
-}
-
-geotrellis.raster.gdal {
-    acceptable-datasets = ["SOURCE", "WARPED"]
-    number-of-attempts  = 1048576
-
-    options {
-        # GDAL_DISABLE_READDIR_ON_OPEN     = "YES"
-        # CPL_VSIL_CURL_ALLOWED_EXTENSIONS = ".tif"
-        # CPL_VSIL_GZIP_WRITE_PROPERTIES   = "NO"
-        GDAL_MAX_DATASET_POOL_SIZE       = "256"
-        GDAL_CACHEMAX                    = "1000" # 500 for 20 nodes cluster
-        # CPL_VSIL_CURL_CHUNK_SIZE         = "1048576"
-        # VRT_SHARED_SOURCE                = "0"
-        # CPL_DEBUG                        = "ON"
-        # GDAL_HTTP_MAX_RETRY              = "10"
     }
 }
